Client: Moved dialog for creating new conversation from messages-page component to a new, new-convo-dialog component file
